Refactoring Code: remove strict mode integration create abstarct class to avoid duplicate code
